Change dinst ID start from 1. 0 becomes an invalid ID. Add load source ID to spec LSQ entry. Memory bypass & dependency & consistency check are both on some granularity: word. Memory consistency & dependency check is based on load source ID. Add header for SC/TSO LSQ L1$ will search LSQ on cache replacement/invalidation.

+++ simu/libmem/CacheArray.h
@@ -22,8 +22,8 @@ public:
   const uint32_t  setNum; // [sizhuo] number of cache sts
 
   const uint32_t  log2Assoc;
-  const uint64_t  log2LineSize;
-  const uint64_t  maskAssoc;
+  const uint32_t  log2LineSize;
+  const uint32_t  maskAssoc;
   const uint32_t  maskSets; // [sizhuo] mask for index calculation
   const uint32_t  log2Sets;
 	const uint32_t  maskBank;
